Updating type attributes for entities

+++ src/main/java/dev/receipt/processor/domain/entity/Points.java
@@ -1,3 +1,3 @@
 package dev.receipt.processor.domain.entity;
 
-public record Points(Integer points) {}
+public record Points(Long points) {}
